Fix issue causing collins to fail on start
Collins was failing to start with the change made to the location of the
profile.yaml file. Moved it back so collins will start up again. Need to
look into why this actually started failing later i'm sure it's nothing
to do with the path and something to do with the base image having a different
file in that location that may have been getting loaded in instead.